Modify typecheck of assign statements:  - Quantify only the free type variables present in the type of the  expression that are NOT free in gamma, after applying the substitution we  obtain in the typechecking of the expression to gamma
Previously, we quantified over the tv's that were free in the type of
the expression and not free in gamma BEFORE applying the substitution,
but this is wrong - the subsitution gives us new information about
gamma. We apply the substitution, and if there STILL remain free type
variables in the expr type, that means they are completely unrelated and
unrestricted by the types in gamma, and we can quantify over them. This
also mirrors the behavior in algorithm W.